HPS-1246: Add ContactListWidget component for displaying a list of contacts
This commit adds a new component, ContactListWidget, to the src/section/crm/customer/contact directory. The ContactListWidget component is responsible for rendering a list of contacts associated with a customer. It fetches the contacts data from the server and displays it in a grid layout. This new component enhances the CRM functionality by providing a visually appealing and user-friendly way to view and interact with customer contacts.