Add flag alias to the project
This commit adds the flag asset to the project by including the path to the flag SVG file in the `vite.config.js` file. The flag asset is located in the `public/asset/svg/flag` directory. This feature enhancement allows the application to display flag images dynamically, improving the visual representation of the content.