HPS-1166: [1st Stage] Add useFilter hook and query tool
The code changes introduce a new `useFilter` hook and a `query-tool` component. The `useFilter` hook provides a filter function that can be used to filter data based on specified parameters. The `query-tool` component utilizes the `useFilter` hook to filter and sort data. This feature enhancement improves the flexibility and efficiency of data filtering and querying in the application.