refactor: Add getMachines thunk for fetching machines with pagination
This code change adds a new `getMachines` thunk to the `machine` slice in `machine.js`. The thunk is responsible for fetching machines with pagination from the server. It sends a GET request to the server with the specified page, pageSize, and isArchived parameters. Upon successful response, the fetched machines are stored in the state using the `getMachinesSuccess` action. In case of an error, the appropriate error message is displayed using the `hasError` action. This enhancement improves the functionality of fetching machines and provides better control over pagination in the application.