HPS-1166: Add useFilter hook and use-filter module
This code change adds the `useFilter` hook and the `use-filter` module to the `hook` directory. The `useFilter` hook provides a filter function that can be used to filter data based on specified parameters. The `use-filter` module exports the `useFilter` hook for easy import and use in other components. This feature enhancement improves the flexibility and efficiency of data filtering in the application.