Update axios import in util.js for authentication
This code change updates the import statement for the axios instance in the `util.js` file. The previous import was using the `axios` object from the `util` module, but it has been changed to import the `axiosInstance` from the `util/axios` module. This update ensures that the correct axios instance is used for making API requests in the authentication module, improving the consistency and reliability of the code.