feat: decoiler icon tooltip indentifier to machine layout
This code change adds a decoiler icon tooltip to the machine layout in the `MachineLayout` component. The tooltip is displayed when the machine model includes any of the decoiler types defined in the `DECOILER_TYPE_ARR` constant. The tooltip provides additional information about the decoiler type and improves the user experience by providing context for the displayed icons.