refactor: Update machine model slice to include decoiler information
This code change updates the machine model slice in the `model.js` file. It adds new state properties `isDecoiler` and `decoilerIcon` to track the decoiler information for each machine model. It also includes new reducer functions `setIsDecoiler`, `setDecoilerIcon`, and `resetDecoilerIcon` to handle the changes to these properties. This update improves the functionality of the machine model slice by allowing the tracking and manipulation of decoiler information.