style: Update tooltip color and text color in global styles
This code change updates the global styles in the `global.jsx` file. It modifies the `GStyledTooltip` component to change the background color of the tooltip and the text color. The background color is updated to the value of the `tooltipcolor` variable, and the text color is set to black. This update improves the visual appearance and readability of tooltips throughout the application.