HPS-1271 :Update StyledFieldGrid background color and border in viewform/style.jsx
This code change updates the background color and border of the StyledFieldGrid component in the viewform/style.jsx file. The background color is modified based on the mode and isMachineView props, and the border is added for the isMachineView prop. This update improves the visual appearance and layout of the StyledFieldGrid component in the view form.