Moved:  BadgeCardMedia component for displaying badge images to Global components
This code change introduces a new component called BadgeCardMedia, which is responsible for displaying badge images in a circular badge overlay on top of a CardMedia component. The component includes functionality for uploading and editing badge images, as well as handling hover events to show an edit icon. This addition enhances the user interface by providing a visually appealing way to showcase badge images.