Moved BadgeCardMedia component to Global components
This code change moves the BadgeCardMedia component from the machine directory to the Global components directory. The BadgeCardMedia component is responsible for displaying badge images in a circular badge overlay on top of a CardMedia component. This reorganization improves code structure and promotes reusability of the BadgeCardMedia component across different sections of the application.