refactor to Thunk: Add getCustomer function to retrieve customer details
This code change adds a new function, getCustomer, to retrieve customer details. The function is implemented as an async action using axiosInstance to make a GET request to the server. Upon successful response, the customer details are dispatched using the getCustomerSuccess action. In case of an error, the hasError action is dispatched and the error is thrown. This enhancement improves the ability to fetch and display specific customer information in the application.