refactor: Update StyledFieldChip component in viewform
This code change updates the StyledFieldChip component in the viewform module. It adds a borderRadius style property to give the chips a rounded appearance. The border property is commented out to remove the border around the chips. Additionally, the backgroundColor property is modified to use the correct shade of grey based on the theme mode. This change improves the visual presentation of the field chips in the form view.