Refax: HowickResources Component
Update the HowickResources component in the src/section/common/howick-resources.jsx file. Add two new props, spacing and isDialog, to customize the grid spacing and determine if the component is used in a dialog. This change improves the flexibility and reusability of the HowickResources component by allowing developers to control the grid spacing and adapt its behavior for different contexts.