🦄 refactor(refax: Update key.js and label.js with new constants ):
This code change updates the key.js and label.js files to include new constants. In key.js, the constants FLEX, FLEX_DIR, and SZ are imported from the flex.js and size.js files respectively. These constants are then added to the KEY object. In label.js, the constants NAVIGATE_TO and PARENT_MACHINE are added to the LABEL object. These changes enhance the functionality and expand the available options in the application.