added city for easy mapping: Add installation site city to machine default values
This code change adds the installationSiteCity property to the machine default values in the default-values.jsx file. It retrieves the city value from the installation site address and assigns it to the installationSiteCity property. This enhancement improves the completeness of the default values by including the installation site city information.