HPS-1278: Add SiteDialog component for managing site information
This code change adds a new component called SiteDialog, which is responsible for managing site information in the application. The SiteDialog component includes fields for entering site details such as name, address, city, region, country, etc. It also displays a map with the location coordinates if valid coordinates are provided. This addition enhances the functionality of the application by allowing users to easily manage and view site information.