HPS-1286: Update Clock component to use main prop for displaying Auckland time
This code change refactors the Clock component to use the main prop for displaying the Auckland time. Instead of always showing the local time, the component now checks if the main prop is true and displays the Auckland time using the new aucklandTime state. This improvement enhances the flexibility of the Clock component by allowing it to display different times based on the main prop value.