HPS-1224: Add local prop to Clock component for displaying local time
The code change adds a new prop, `local`, to the Clock component. This prop allows the component to display the local time instead of the Auckland time. This enhancement provides flexibility to the Clock component by allowing it to show different times based on the value of the `local` prop.