HPS-1273: Add customer overview button to CustomerDialog component
This code change adds a new button to the CustomerDialog component that allows users to navigate to the customer overview page. When clicked, the button triggers the handleCustomerOverview function, which sets the customer dialog to false and navigates to the customer overview page. This feature enhances the user experience by providing easy access to detailed customer information.