HPS-1277: Refactor ContactListWidget component
This code change refactors the ContactListWidget component by removing the commented-out code for opening the contact view page and adding a new prop, handleContactDialog, to handle the click event for viewing contact details. This refactor improves the readability and maintainability of the codebase by removing unused code and enhancing the functionality of the component.