Update machineDefaultValues function to include installation site name and address
This code change updates the machineDefaultValues function in the default-values.jsx file to include the installation site name and address. The installationSiteName and installationSiteStreet properties are added to the returned object, using the values from the machineData object. This refactor improves the completeness and accuracy of the default values for the machine form.