Update disabled state colors in ACTION object
This code change updates the disabled state colors in the ACTION object of the palette.js file. The disabled color is changed to alpha(GREY[400], 1) and the disabledText color is changed to alpha(GREY[700], 0.46). Additionally, the disabledBackground color is changed to alpha(GREY[700], 1). This refactor improves the consistency and visual appearance of the disabled state in the application.