HPS-1297: [Housekeeping] Add new web icons for eye and eye-off
This code change adds new web icons for the eye and eye-off functionalities. The ICON_WEB object in the icon-directory.js file is updated to include the EYE and EYE_OFF icons, and the ICON_WEB_NAME object is also updated accordingly. This feature enhancement improves the user interface by providing visually appealing icons for the eye and eye-off actions.