Add alongTab prop to BackButton component
This code change adds a new prop called alongTab to the BackButton component. When alongTab is set to true, the BackButton component will be rendered as an IconTooltip component, providing a tooltip with the label "Go Back". This change enhances the user experience by allowing more flexibility in rendering the BackButton component in different scenarios.