HPS-1310: Update login layout and style
This code change updates the login layout and style to improve the user experience and align with the new branding. The login layout now displays the app branding instead of the app branch, and the background color is adjusted based on the theme mode. Additionally, the typography variant for the login title is changed to uppercase for better visibility. These changes enhance the visual appeal and consistency of the login page.