add customerTab state to customer slice
This code change adds the customerTab state to the customer slice in the CRM module. The customerTab state is used to keep track of the currently selected tab in the customer section. This enhancement improves the functionality of the CRM module by allowing users to easily switch between different tabs within the customer section.