refactor: Add flag parameter to getCustomer API call
This code change adds a flag parameter to the getCustomer API call in the customer.js file. The flag parameter is set to 'basic' and is used to retrieve basic customer information. This enhancement improves the functionality of the API call and ensures that only the necessary data is fetched.