HPS-1312: site default values and site card components
This code change refactors the site default values and site card components. The `default-values.jsx` file is updated to use the `name` property instead of `firstName` and `lastName` for the site's name. The `site-card.jsx` file is modified to display the city and country of the site's address instead of the contact's title and email. These changes improve the consistency and clarity of the codebase.