HWKSC-1321: Add Customer Support Tickets page
This commit adds a new file `tickets.jsx` under the `src/page/crm/support` directory. The file contains a functional component `CustomerSupportTickets` that renders the `CustomerLayout` component with the `TicketsTab` component inside. This addition introduces a new page for managing customer support tickets in the CRM section of the application.